requirements.txt
# Web Framework & Server
fastapi
uvicorn

# Streaming & Messaging
kafka-python

# Big Data & Processing
pyspark

# Cloud & Storage
google-cloud-firestore

# Machine Learning & NLP
transformers
torch

# API & Web Scraping
requests
praw

# Environment Variables & Database
python-dotenv
psycopg2-binary
sqlalchemy  # Added for PostgreSQL ORM support (optional)
pandas